+++ gwcelery/tests/test_sentry.py
@@ -21,10 +21,11 @@ def netrcfile(monkeypatch, tmpdir):
 @patch('sentry_sdk.init')
 @patch('sentry_sdk.integrations.celery.CeleryIntegration')
 @patch('sentry_sdk.integrations.flask.FlaskIntegration')
+@patch('sentry_sdk.integrations.redis.RedisIntegration')
 @patch('sentry_sdk.integrations.tornado.TornadoIntegration')
-def test_sentry_configure(mock_tornado_integration, mock_flask_integration,
-                          mock_celery_integration, mock_sdk_init, netrcfile,
-                          caplog):
+def test_sentry_configure(mock_tornado_integration, mock_redis_integration,
+                          mock_flask_integration, mock_celery_integration,
+                          mock_sdk_init, netrcfile, caplog):
     caplog.set_level(logging.ERROR)
     sentry.configure()
     record, = caplog.records
@@ -45,4 +46,5 @@ def test_sentry_configure(mock_tornado_integration, mock_flask_integration,
         integrations=[
             mock_celery_integration.return_value,
             mock_flask_integration.return_value,
+            mock_redis_integration.return_value,
             mock_tornado_integration.return_value])